pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-stock-market-analysis"
version = "1.0.0"
description = "Stock market analysis application using AlphaVantage API with MCP integration"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Stock Analysis Team"}
]
dependencies = [
    "anthropic>=0.15.0",
    "mcp[cli]>=0.2.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
    "pydantic>=2.0.0",
    "requests>=2.30.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.24.0",
    "plotly>=5.14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
